generic, doesnt need to readium css
