+++ r2-shared/src/main/java/org/readium/r2/shared/Publication.kt
@@ -3,8 +3,6 @@ package org.readium.r2.shared
 import android.content.SharedPreferences
 import org.json.JSONArray
 import org.json.JSONObject
-import org.readium.r2.shared.UserSettings.UserSetting
-import org.readium.r2.shared.UserSettings.UserSettings
 import java.io.Serializable
 import java.net.URL
 
@@ -76,14 +74,11 @@ class Publication() : Serializable {
     var otherLinks: MutableList<Link> = mutableListOf()
     var internalData: MutableMap<String, String> = mutableMapOf()
     //var manifestDictionnary: Map<String, Any> = mapOf()
+
     var coverLink: Link?  = null
         get() = linkWithRel("cover")
 
-    lateinit var userSettings: UserSettings
-
-    constructor(sharedPreferences: SharedPreferences) : this() {
-        userSettings = UserSettings(sharedPreferences)
-    }
+    var userSettings = UserSettings()
 
     fun baseUrl() : URL? {
         val selfLink = linkWithRel("self")
