--- r2-shared/src/main/java/org/readium/r2/shared/UserSettings/Appearance.kt
@@ -1,7 +1,0 @@
-package org.readium.r2.shared.UserSettings
-
-class Appearance(var value: String) : Enumeratable(APPEARANCE_REF, APPEARANCE_NAME, value) {
-
-    override val values = listOf("readium-default-on", "readium-sepia-on", "readium-night-on")
-
-}
